python/requirements.txt
fastapi
uvicorn
tensorflow
numpy
opencv-python-headless
mediapipe
pillow
google-cloud-storage
python-dotenv